Convert more components and mixins to Vue single file components.
Several issues still remaining:

 - some deps (e.g. CodeMirror) not working via imports yet
 - Vue filters etc still in raw JS
