+++ modules/admin/app/assets/js/components/mixin-preview.vue
@@ -1,0 +1,15 @@
+<script>
+export default {
+  data: function() {
+    return {
+      tab: 'preview'
+    }
+  },
+  methods: {
+    showPreview: function (file) {
+      this.previewing = file;
+      this.tab = 'preview';
+    },
+  }
+}
+</script>
