+++ modules/admin/app/assets/js/components/mixin-two-panel.vue
@@ -1,0 +1,15 @@
+<script>
+
+export default {
+    data: function() {
+    return {
+      panelSize: null,
+    }
+  },
+  methods: {
+    setPanelSize: function (arbitrarySize) {
+      this.panelSize = arbitrarySize;
+    }
+  }
+}
+</script>
