+++ modules/admin/app/assets/js/components/modal-info.vue
