+++ modules/admin/app/assets/js/components/panel-file-preview.vue
@@ -1,0 +1,8 @@
+<script>
+
+import MixinPreviewPanel from "./mixin-preview-panel"
+
+export default {
+  mixins: [MixinPreviewPanel]
+}
+</script>
