+++ modules/admin/app/assets/js/components/rs-manager.vue
@@ -5,13 +5,22 @@ import ValidateButton from './validate-button';
 import DeleteButton from './delete-button';
 import FilesTable from './files-table';
 import DragHandle from './drag-handle';
-import InfoModal from './info-modal';
+import PanelFilePreview from './panel-file-preview';
+import ModalInfo from './modal-info';
 import LogWindow from './log-window';
 import RsConfigModal from './rs-config-modal';
 
+import MixinTwoPanel from './mixin-two-panel';
+import MixinValidator from './mixin-validator';
+import MixinPreview from './mixin-preview';
+import MixinError from './mixin-error';
+import MixinStage from './mixin-stage';
+import MixinUtil from './mixin-util';
+
+
 export default {
-  components: {FilterControl, FilesTable, LogWindow, DragHandle, InfoModal, ValidateButton, DeleteButton, RsConfigModal},
-  mixins: [stageMixin, twoPanelMixin, previewMixin, validatorMixin, errorMixin, utilMixin],
+  components: {FilterControl, FilesTable, LogWindow, DragHandle, ModalInfo, PanelFilePreview, ValidateButton, DeleteButton, RsConfigModal},
+  mixins: [MixinStage, MixinTwoPanel, MixinPreview, MixinValidator, MixinError, MixinUtil],
   props: {
     fileStage: String,
     config: Object,
@@ -133,7 +142,7 @@ export default {
           v-on:error="showError"
           v-on:close="showOptions = false"/>
 
-      <info-modal v-if="fileInfo !== null" v-bind:file-info="fileInfo" v-on:close="fileInfo = null"/>
+      <modal-info v-if="fileInfo !== null" v-bind:file-info="fileInfo" v-on:close="fileInfo = null"/>
     </div>
 
     <div id="rs-panel-container" class="panel-container">
@@ -199,7 +208,7 @@ export default {
 
         <div class="status-panels">
           <div class="status-panel" v-show="tab === 'preview'">
-            <preview v-bind:dataset-id="datasetId"
+            <panel-file-preview v-bind:dataset-id="datasetId"
                      v-bind:file-stage="fileStage"
                      v-bind:previewing="previewing"
                      v-bind:panel-size="panelSize"
