+++ modules/admin/app/assets/js/components/xslt-editor.vue
@@ -1,4 +1,7 @@
 <script>
+
+// import CodeMirror from 'codemirror';
+
 export default {
   props: {
     value: String,
