--- modules/admin/app/assets/js/datamanager.js
@@ -1,11 +1,0 @@
-
-import DatamanagerApp from './datamanager-app';
-
-Vue.config.productionTip = false
-
-/* eslint-disable no-new */
-new Vue({
-  el: '#app',
-  template: '<DatamanagerApp/>',
-  components: { DatamanagerApp }
-})
