+++ modules/admin/app/views/admin/repository/datamanager.scala.html
@@ -81,8 +81,8 @@
         </script>
 
         <script src="@controllers.admin.routes.AdminAssets.versioned("js/lib/vue.js")"></script>
-        <script src="@controllers.admin.routes.AdminAssets.versioned("js/lib/Sortable.js")"></script>
-        <script src="@controllers.admin.routes.AdminAssets.versioned("js/lib/vuedraggable.umd.js")"></script>
+@*        <script src="@controllers.admin.routes.AdminAssets.versioned("js/lib/Sortable.js")"></script>*@
+@*        <script src="@controllers.admin.routes.AdminAssets.versioned("js/lib/vuedraggable.umd.js")"></script>*@
         <script src="@controllers.admin.routes.AdminAssets.versioned("js/lib/axios.js")"></script>
         <script src="@controllers.admin.routes.AdminAssets.versioned("js/lib/lodash.js")"></script>
         <script src="@controllers.admin.routes.AdminAssets.versioned("js/lib/luxon.js")"></script>
